changed class based files + added Arduino_LSM9DS1.h library to try and increase gyro output rate (unsuccessful so far).

+++ C Files/calc_complementary_angle/calculate_tilt_angle_functions.h
@@ -1,7 +1,7 @@
 #ifndef CALCULATE_TILT_ANGLE_FUNCTIONS_H
 #define CALCULATE_TILT_ANGLE_FUNCTIONS_H
 
-float calc_acc_tilt_angle(int ya, int yz);
+float calc_acc_tilt_angle(float ya, float yz);
 float calc_gyro_tilt_angle(float last_gyro_angle, float last_timestamp, float current_timestamp, float zg);
 
 #endif
